Added flag prefab
Added a flag prefab, also some capture the flag logic in FlagController.cs.
I tried to add the player into my scene, but couldn't get everything to work right. When this is ready to be merged into main, I'd like to have somebody walk me through how to add the player prefab to my scene so that it all works and there isn't any errors