Elyes Player Merge Resolution
I hand merged all Elyes files from his pull request. I had to "re-prefab" the PlayerArmature, there was a few unsaved changes in it.
I tested it all out on my scene, and it all works good with no errors.
Only problem is that the Elyes scene is pretty borked at a Unity level, so Elyes you are gonna have to make a new scene.