Fixed CTF capture player movement bug
Fixed issue where scoring the flag propelled the player forward.
This was due to having multiple colliders on the flag (parent collider and orb collider) which would score the flag multiple times.
Cleaning this up allowed for the removal of the Base scripts, because they no longer need to watch for orb collisions. Everything is handled within the flag